Moved middleware approppriately
