+++ src/controllers/httpHandler.ts
@@ -1,0 +1,8 @@
+import winston = require("winston");
+
+export class DefaultHTTPHandler {
+  logger: winston.Logger;
+  constructor(logger: winston.Logger) {
+    this.logger = logger;
+  }
+}
