+++ src/controllers/user/routes.ts
@@ -1,41 +1,87 @@
 import { Router } from "restify-router";
-import { authentication } from "./authMiddleware";
-import { EventTypes } from "../../types/events";
+import admin = require("../../utils/firebase");
+import * as errors from "restify-errors";
 import logger = require("../../utils/logger");
-import {
-  setUserProperties,
-  trackUserEvent,
-  incrementUserProperties,
-} from "../../utils/mixpanel";
+import { UserService } from "./service";
+import { DefaultHTTPHandler } from "../httpHandler";
+import winston = require("winston");
+import { Next, Response } from "restify";
+import testUserStub from "../../tests/helpers/stubs/testUserStub";
+import { ErrorTypes } from "../../types/errors";
+import { RequestWithContext } from "../../types/restify";
 
-export class UserHandler {
-  SetupRoutes(): Router {
-    let PingRouter = new Router();
+export class UserHTTPHandler extends DefaultHTTPHandler {
+  userService: UserService;
 
-    PingRouter.get("/ping", authentication, authorizedPingAndGetOKResponse);
-    PingRouter.get("/freeping", unAuthorizedPingAndGetOKResponse);
+  constructor(logger: winston.Logger, userService: UserService) {
+    super(logger);
+    this.userService = userService;
+  }
+
+  public SetupRoutes(): Router {
+    const PingRouter = new Router();
+    PingRouter.get(
+      "/ping",
+      this.AuthMiddleware,
+      this.authorizedPingAndGetOKResponse
+    );
+    PingRouter.get("/freeping", this.unAuthorizedPingAndGetOKResponse);
     return PingRouter;
   }
-}
 
-async function authorizedPingAndGetOKResponse(req, res, next) {
-  res.send({ ping: "Authorized OK" });
-  const user = req.get("user");
-  await setUserProperties(user.uid, {
-    email_verified: true,
-    money: 200 + Math.random() * 1000,
-  });
+  public async AuthMiddleware(
+    req: RequestWithContext,
+    res: Response,
+    next: Next
+  ) {
+    let decodeValue;
 
-  trackUserEvent(EventTypes.Ping, user.uid);
+    const token = req.headers?.authorization?.split(" ")[1];
 
-  incrementUserProperties(user.uid, { money: 1000 });
-  return next();
-}
+    if (!token) {
+      return next(
+        res.send(
+          new errors.UnauthorizedError(
+            { statusCode: 401 },
+            "Bearer Token is missing"
+          )
+        )
+      );
+    }
 
-async function unAuthorizedPingAndGetOKResponse(req, res, next) {
-  res.send({ ping: "Non-Authorized OK" });
+    try {
+      if (process.env.NODE_ENV !== "test") {
+        decodeValue = await admin.auth().verifyIdToken(token);
+      } else {
+        decodeValue = testUserStub.testUser;
+      }
 
-  logger.info("response sent");
+      if (decodeValue) {
+        req.set("user", decodeValue);
+        return next();
+      }
+      next(new errors.UnauthorizedError("Unauthorized"));
+    } catch (err) {
+      if (err.code == ErrorTypes.TokenExpiredError) {
+        return next(new errors.RequestExpiredError(err.message));
+      }
+      if (err.code == ErrorTypes.TokenRevokedError) {
+        return next(new errors.UnauthorizedError(err.message));
+      }
+      next(new errors.InternalServerError(err));
+    }
+  }
 
-  return next();
+  protected async authorizedPingAndGetOKResponse(req, res, next) {
+    const pingResult = this.userService.ping(true);
+    res.send(pingResult);
+    return next();
+  }
+
+  protected async unAuthorizedPingAndGetOKResponse(req, res, next) {
+    const pingResult = this.userService.ping(false);
+    res.send(pingResult);
+    this.logger.info("response sent");
+    return next();
+  }
 }
