+++ src/controllers/user/serviceSchema.ts
@@ -1,0 +1,7 @@
+interface Service {
+  ping(auth: boolean): GenericObject<string>;
+}
+
+type GenericObject<m = any> = {
+  [key: string]: m;
+};
